requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg[binary]==3.2.9
python-dotenv==1.0.1
drf-spectacular==0.27.2
